"Added scrolling text effect on page 5 using GSAP and modified styles for page 4 and 5"
This commit includes changes to the HTML and CSS files for the "GSAP Main Project". In the HTML file, the content within the page5 div has been updated to include four div elements with scrolling text. These text elements are implemented using the data-scroll and data-scroll-speed attributes. In the CSS file, the styles for page4-blue-div have been updated to use a lowercase hexadecimal color code, and new styles have been added for the scrolling text elements in page5. These styles include the animation properties for the scrolling effect. Additionally, the blue-div-elem h4 and p styles have been updated for better readability.